move test function into more appropriate test file
